Move contact details rendering inside WPContactForm
It was previously in an injected function because the form wrapper was
in a separate package, but that is no longer the case.
